feature:1.对接聊天室消息

+++ module_chat/src/main/java/arch/cayenne/module/chat/ui/fragment/ChatBaseFragment.kt
@@ -267,7 +267,7 @@ abstract class ChatBaseFragment : BaseFragment<ChatHomeViewModel, FragmentLiveCh
             sendText()
         }
         mViewModel.atLiveData.observe(viewLifecycleOwner) {
-            chatAtHelper.addAtMentionSpan(it.userName, ChatRefUser(it.uid.toLong(),it.userName,it.avatarId,it.replaceUserName))
+            chatAtHelper.addAtMentionSpan(it.userName, ChatRefUser(it.uid,it.userName,it.avatarId,it.replaceUserName))
             SoftKeyBoardAnim.etAnimWhenEtContentChange(
                 mBinding,
                 mViewModel.currentKeyBoardType,
